Unify look and feel buttons in search result screen and search helper
